+++ Module.php
@@ -2,6 +2,15 @@
 
 namespace humhub\modules\socialshare;
 
+use yii\helpers\Url;
+
 class Module extends \humhub\components\Module
 {
+    public $resourcesPath = 'resources';
+
+    public function getConfigUrl()
+    {
+        return Url::to(['/socialshare/admin']);
+    }
+
 }