+++ assets/Assets.php
@@ -6,12 +6,19 @@ use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
+    /**
+     * @inheritdoc
+     */
     public $sourcePath = '@socialshare/resources';
 
     /**
      * @inheritdoc
      */
+    public $js = [
+        'js/humhub.socialshare.ShareLink.js'
+    ];
+
     public $publishOptions = [
-        'forceCopy' => 'false',
+        'forceCopy' => 'false'
     ];
 }