+++ docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.2.0 (TBA)
+-----------------------
+- Enh: #6 Configurable Social Share Options
+- Chg: Twitter to X
+- Enh: Full Refactor of module
+
 1.1.0 (August 27, 2025)
 -----------------------
 - Fix #10: Update module resources path
@@ -20,3 +26,7 @@ Changelog
 1.0.2 (24 August, 2018)
 -----------------------
 - Fix: PHP 7.2 compatibility
+
+
+
+